Move inner/static classes and interfaces from `InMemoryTransportFactory` to an upper level.
