+++ server/src/main/java/io/spine/server/integration/memory/AbstractInMemoryChannel.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package io.spine.server.integration.memory;
+
+import io.spine.server.integration.ExternalMessageClass;
+import io.spine.server.integration.MessageChannel;
+
+/**
+ * An abstract base for in-memory {@linkplain MessageChannel message channels}.
+ *
+ * @author Alex Tymchenko
+ */
+abstract class AbstractInMemoryChannel implements MessageChannel {
+
+    /**
+     * Message class that defines what messages travel through this channel.
+     */
+    private final ExternalMessageClass messageClass;
+
+    AbstractInMemoryChannel(ExternalMessageClass messageClass) {
+        this.messageClass = messageClass;
+    }
+
+    @SuppressWarnings("NoopMethodInAbstractClass")  // See the method body for the explanation.
+    @Override
+    public void close() throws Exception {
+        // There is nothing to close in the in-memory local channel implementation.
+    }
+
+    @Override
+    public ExternalMessageClass getMessageClass() {
+        return messageClass;
+    }
+}
